Bump to 1.2 and describe new syntax
This is because I wanted to move the whole project in to it's own package and I didn't want to fire up Matlab when just importing the package. Therefore, I added a MatlabInstance management class to get the instance explicitly - which seemed cleaner than the init method I had before.